Fix MIME type detection on Linux
On Linux it seems like JavaScript's `type` attribute for .wav files isn't `audio/wav` but instead `audio/x-wav`. The check is inside the deployed site according to Firefox's debugger in `(index)` (the page itself), line 266.   
I changed it in this pull request to detect both `audio/wav` and `audio/x-wav` as a .wav file. (the contents are the same, the OS just assigns a different MIME type to it)   

Also, I made it not just silently ignore the file and only show the error in console, it now instead popps up saying "Please select a .wav file!".