Added Lp calculation of best peaks and added additional annotation for stat window.
This is currently very similar to the standard deviation and serves as a proof of concept that the Lp norm works, and that it can be applied to a vector of any size.
Can be improved by using a user-provided p.
Can be improved by allowing user to switch between estimated diff mean and target diff mean (as calculated from bpm_target).

If the user provides:
	bpm_target
	A window of beats
Then the Lp norm can be applied to calculate a long-term measure across the window of how well it follows the bpm_target.